+++ lib/debug_logging/class_notifier.rb
@@ -14,7 +14,7 @@ module DebugLogging
         )
         original_method = method(method_to_notify)
         (class << self; self; end).class_eval do
-          define_method(method_to_notify) do |*args, &block|
+          define_method(method_to_notify) do |*args, **kwargs, &block|
             config_proxy = DebugLogging::Util.config_proxy_finder(
               scope: self,
               config_opts: method_config_opts,
@@ -33,19 +33,15 @@ module DebugLogging
             ActiveSupport::Notifications.instrument(
               DebugLogging::ArgumentPrinter.debug_event_name_to_s(method_to_notify: method_to_notify),
               {
-                debug_args: args,
+                debug_args: kwargs.empty? ? args : args + [kwargs],
                 config_proxy: config_proxy,
                 **paydirt,
               },
             ) do
-              if args.size == 1 && (harsh = args[0]) && harsh.is_a?(Hash)
-                original_method.call(**harsh, &block)
-              else
-                original_method.call(*args, &block)
-              end
+              original_method.call(*args, **kwargs, &block)
             rescue StandardError => e
               if config_proxy.error_handler_proc
-                config_proxy.error_handler_proc.call(config_proxy, e, self, method_to_notify, args)
+                config_proxy.error_handler_proc.call(config_proxy, e, self, method_to_notify, *args, **kwargs)
               else
                 raise e
               end